+++ test/browser/features/fixtures/packages/svelte-kit/src/routes/contact/[contactId]/+page.svelte
@@ -1,0 +1,11 @@
+<script>
+    import { page } from '$app/state';
+
+    const title = `Contact ${page.data.contactId}`;
+</script>
+
+<svelte:head>
+	<title>{title}</title>
+</svelte:head>
+
+<h1>{title}</h1>